Remove reference from setting to device
Now user code (device module implementations) are responsible for
keeping a reference to the device through the setting implData.

This decreases typing for the majority of settings (and increases typing
for settings that are parameterized over more than the device identity),
but the main goal is to make settings an independently reusable object,
so that things other than devices can also have settings.