Document some API functions
Most of the functions with non-trivial memory management semantics are
documented, as well as those that definitely need to change in the near
future.

Added OSc_MAX_STR_SIZE and OScDev_MAX_STR_SIZE so that we don't need to
keep writing OSc_MAX_STR_LEN + 1.

Added a null check for OSc_Device_SetLogFunc().